Added LedgerDialog.tsx and updated LockGold dialogs
