+++ components/ledger/celoGold/lock/LockGoldConfirm.tsx
@@ -14,16 +14,6 @@ import Ledger from '../../Ledger'
 
 const useStyles = makeStyles({
 
-    title: {
-        display: "block",
-        textAlign: "center",
-        paddingTop: "0.5rem",
-        paddingBottom: "0.7rem",
-    },
-
-    dialogTitle: {
-        padding: "1rem 1rem 0rem 1rem",
-    },
 
     dialogContent: {
         display: "flex",
@@ -40,14 +30,6 @@ const useStyles = makeStyles({
 
     item: {
         justifyContent: "center",
-    },
-
-
-    iconButtonRight: {
-        float: "right"
-    },
-    iconButtonLeft: {
-        float: 'left'
     },
 
     lockGoldMessage: {
@@ -67,26 +49,16 @@ const useStyles = makeStyles({
         paddingTop: "0.5rem",
         paddingBottom: "0.5rem",
     },
-    errorMessage: {
-        color: "red",
-        textAlign: "center",
-        //paddingBottom: "1rem"
-    },
 
 });
 
-type LockGoldConfirmProps = { isOpen: boolean, amount: string, pageAddress?: string };
+type LockGoldConfirmProps = { amount: string };
 
 
-const LockGoldConfirm = ({ isOpen, amount, pageAddress }: LockGoldConfirmProps) => {
+const LockGoldConfirm = ({ amount }: LockGoldConfirmProps) => {
 
     const classes = useStyles();
-    const [open, setOpen] = React.useState(isOpen);
     const [currentUser, setCurrentUser] = React.useState('');
-    const [previousDialog, setPreviousDialog] = React.useState(false)
-    const [nextDialog, setNextDialog] = React.useState(false);
-    const [errorMessage, setErrorMessage] = React.useState('');
-    const [currentAddress, setCurrentAddress] = React.useState(pageAddress || '');
     const [lockAmount, setLockAmount] = React.useState(amount);
 
 
@@ -95,169 +67,113 @@ const LockGoldConfirm = ({ isOpen, amount, pageAddress }: LockGoldConfirmProps) 
         //@ts-ignore
         setCurrentUser(localUser)
     });
-    const handleClose = () => {
-        setOpen(false);
-    };
-
-    const handlePreviousDialog = () => {
-        setOpen(false)
-        setPreviousDialog(true);
-    };
-
-    const handleNextDialog = () => {
-        setOpen(false)
-        setNextDialog(true)
-    }
 
     return (
         <>
-            <Dialog
-                open={open}
-                onClose={handleClose}
-                aria-labelledby="ledger-dialog-lock-gold-confirm"
-                //fullWidth
-                maxWidth="sm"
-            >
-                <DialogTitle id="ledger-lock-gold-confirm" className={classes.dialogTitle}>
-                    <Grid container className={classes.item}>
-                        <Grid item xs={1}>
-                            <IconButton
-                                aria-label="Return"
-                                className={classes.iconButtonLeft}
-                                onClick={handlePreviousDialog}
-                            >
-                                <img src="/images/last.svg" color="textPrimary" />
-                            </IconButton>
-                        </Grid>
-                        <Grid item xs={10}>
-                            <Typography variant="h6" color="textPrimary" noWrap className={classes.title}>
-                                Lock CELO
-              </Typography>
-                        </Grid>
-                        <Grid item xs={1}>
-                            <IconButton
-                                aria-label="Close"
-                                className={classes.iconButtonRight}
-                                onClick={handleClose}
-                            >
-                                <img src="/images/cross.svg" color="textPrimary" />
-                            </IconButton>
-                        </Grid>
-                    </Grid>
-                </DialogTitle>
-
-                <DialogContent >
-                    <Grid container spacing={1} >
-                        <DialogContentText id="ledger-lock-gold-confirm-content" className={classes.dialog}>
-                            <Grid container className={classes.dialogContent}>
-                                <Grid item xs={12}>
-                                    <Typography variant="body2" noWrap={false} color="textPrimary" gutterBottom>
-                                        You are going to lock {lockAmount} CELO, it that's correct, please
-                  sign in your ledger device.
-                </Typography>
-                                </Grid>
-                                <Grid item xs={12}>
-                                    <Divider className={classes.divider} />
-                                </Grid>
-                                <Grid container spacing={1} className={classes.item}>
-                                    <Grid item xs={6}>
-                                        <Typography
-                                            variant="body2"
-                                            noWrap
-                                            className={classes.alignLeft}
-                                            align="left"
-                                            color="textPrimary"
-                                        >
-                                            Account
-                  </Typography>
-                                    </Grid>
-                                    <Grid item xs={6}>
-                                        <Typography
-                                            variant="body2"
-                                            noWrap
-                                            className={classes.alignRight}
-                                            align="right"
-                                            color="textPrimary"
-                                        >
-                                            {currentUser}
-                                        </Typography>
-                                    </Grid>
-                                </Grid>
-                                <Grid item xs={12}>
-                                    <Divider className={classes.divider} />
-                                </Grid>
-                                <Grid container spacing={1} className={classes.item}>
-                                    <Grid item xs={6}>
-                                        <Typography
-                                            variant="body2"
-                                            noWrap
-                                            className={classes.alignLeft}
-                                            align="left"
-                                            color="textPrimary"
-                                        >
-                                            Lock Amount
-                  </Typography>
-                                    </Grid>
-                                    <Grid item xs={6}>
-                                        <Typography
-                                            variant="body2"
-                                            noWrap
-                                            className={classes.alignRight}
-                                            align="right"
-                                            color="textPrimary"
-                                        >
-                                            {lockAmount} CELO
-                  </Typography>
-                                    </Grid>
-                                </Grid>
-                                <Grid item xs={12}>
-                                    <Divider className={classes.divider} />
-                                </Grid>
-                                <Grid container spacing={1} className={classes.item}>
-                                    <Grid item xs={6}>
-                                        <Typography
-                                            variant="body2"
-                                            noWrap
-                                            className={classes.alignLeft}
-                                            align="left"
-                                            color="textPrimary"
-                                        >
-                                            Tx Fee
-                  </Typography>
-                                    </Grid>
-                                    <Grid item xs={6}>
-                                        <Typography
-                                            variant="body2"
-                                            noWrap
-                                            className={classes.alignRight}
-                                            align="right"
-                                            color="textPrimary"
-                                        >
-                                            {"0.00001"} CELO
-                  </Typography>
-                                    </Grid>
-                                </Grid>
-                                <Grid item xs={12}>
-                                    <Divider className={classes.divider} />
-                                </Grid>
-                                <Grid item xs={12} className={classes.lockGoldMessage}>
-                                    <Typography variant="h6" noWrap align="center" color="textPrimary" >
-                                        Please sign in your ledger device...
-                </Typography>
-                                </Grid>
-                                {errorMessage ?
-                                    <Grid item xs={12} className={classes.errorMessage}>
-                                        <Typography variant="body2">
-                                            {errorMessage}
-                                        </Typography>
-                                    </Grid> : null}
+            <DialogContent >
+                <Grid container spacing={1} >
+                    <DialogContentText id="ledger-lock-gold-confirm" className={classes.dialog}>
+                        <Grid container className={classes.dialogContent}>
+                            <Grid item xs={12}>
+                                <Typography variant="body2" noWrap={false} color="textPrimary" gutterBottom>
+                                    You are going to lock {lockAmount} CELO, it that's correct, please
+                                    sign in your ledger device.
+                                </Typography>
                             </Grid>
-                        </DialogContentText>
-                    </Grid>
-                </DialogContent>
-            </Dialog>
-            { previousDialog ? <LockGold isOpen={previousDialog} showButton={false} pageAddress={currentAddress} /> : null}
-            {nextDialog ? <LockGoldSuccess isOpen={nextDialog} /> : null}
+                            <Grid item xs={12}>
+                                <Divider className={classes.divider} />
+                            </Grid>
+                            <Grid container spacing={1} className={classes.item}>
+                                <Grid item xs={6}>
+                                    <Typography
+                                        variant="body2"
+                                        noWrap
+                                        className={classes.alignLeft}
+                                        align="left"
+                                        color="textPrimary"
+                                    >
+                                        Account
+                                    </Typography>
+                                </Grid>
+                                <Grid item xs={6}>
+                                    <Typography
+                                        variant="body2"
+                                        noWrap
+                                        className={classes.alignRight}
+                                        align="right"
+                                        color="textPrimary"
+                                    >
+                                        {currentUser}
+                                    </Typography>
+                                </Grid>
+                            </Grid>
+                            <Grid item xs={12}>
+                                <Divider className={classes.divider} />
+                            </Grid>
+                            <Grid container spacing={1} className={classes.item}>
+                                <Grid item xs={6}>
+                                    <Typography
+                                        variant="body2"
+                                        noWrap
+                                        className={classes.alignLeft}
+                                        align="left"
+                                        color="textPrimary"
+                                    >
+                                        Lock Amount
+                                    </Typography>
+                                </Grid>
+                                <Grid item xs={6}>
+                                    <Typography
+                                        variant="body2"
+                                        noWrap
+                                        className={classes.alignRight}
+                                        align="right"
+                                        color="textPrimary"
+                                    >
+                                        {lockAmount} CELO
+                                    </Typography>
+                                </Grid>
+                            </Grid>
+                            <Grid item xs={12}>
+                                <Divider className={classes.divider} />
+                            </Grid>
+                            <Grid container spacing={1} className={classes.item}>
+                                <Grid item xs={6}>
+                                    <Typography
+                                        variant="body2"
+                                        noWrap
+                                        className={classes.alignLeft}
+                                        align="left"
+                                        color="textPrimary"
+                                    >
+                                        Tx Fee
+                                    </Typography>
+                                </Grid>
+                                <Grid item xs={6}>
+                                    <Typography
+                                        variant="body2"
+                                        noWrap
+                                        className={classes.alignRight}
+                                        align="right"
+                                        color="textPrimary"
+                                    >
+                                        {"UNKNOWN"} CELO
+                                    </Typography>
+                                </Grid>
+                            </Grid>
+                            <Grid item xs={12}>
+                                <Divider className={classes.divider} />
+                            </Grid>
+                            <Grid item xs={12} className={classes.lockGoldMessage}>
+                                <Typography variant="h6" noWrap align="center" color="textPrimary" >
+                                    Please sign in your ledger device...
+                </Typography>
+                            </Grid>
+                        </Grid>
+                    </DialogContentText>
+                </Grid>
+            </DialogContent>
+
         </>
     );
 };
